separate file tests
To not have 2 `runTests` in same file.
